add Cab in expenses

+++ app/src/main/java/com/example/moneytracker/rest/RestService.java
@@ -45,7 +45,7 @@ public class RestService {
         return restClient.getSyncUserApi().syncExpense(DataBaseApp.getGoogleToken(context), data, DataBaseApp.getAuthKey());
     }
 
-    public UserExpenseModel addExpense(int sum, String comment, int categoryId, String trDate, String gToken) {
+    public UserExpenseModel addExpense(double sum, String comment, int categoryId, String trDate, String gToken) {
         return restClient.getExpenseUserApi().addExpense(sum, comment, categoryId, trDate, gToken, DataBaseApp.getAuthKey());
     }
